feat: API 수정 및 토큰 적용
