+++ server/package.json
@@ -11,7 +11,9 @@
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.21.0",
+    "jsonwebtoken": "^9.0.2",
     "mariadb": "^3.3.2",
     "reflect-metadata": "^0.2.2",
     "sse": "^0.0.8",
@@ -21,6 +23,7 @@
   "devDependencies": {
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
+    "@types/jsonwebtoken": "^9.0.7",
     "@types/sse": "^0.0.0",
     "@types/swagger-jsdoc": "^6.0.4",
     "swagger-ui-express": "^5.0.1",
