+++ server/src/config/db.ts
@@ -1,9 +1,12 @@
 import mariadb from 'mariadb';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 const pool = mariadb.createPool({
   host: 'localhost',
-  user: 'root',
-  password: 'root',
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
   database: 'TODO_APP',
   connectionLimit: 5,
   timezone: '+09:00',
