+++ server/src/middlewares/middleware.ts
@@ -4,6 +4,7 @@ import { TodoService } from '../services/todoService';
 import { AuthService } from '../services/authService';
 import { CategoryService } from '../services/categoryService';
 import { ReminderService } from '../services/reminderService';
+import { UserService } from '../services/userService';
 
 const connectionPromise = getConnection();
 
@@ -14,6 +15,7 @@ export const Middleware = async (req: Request, _res: Response, next: NextFunctio
     req.authService = new AuthService(conn);
     req.categoryService = new CategoryService(conn);
     req.reminderService = new ReminderService(conn);
+    req.userService = new UserService(conn);
     req.middlewareProcessed = true;
     next();
   } catch (error) {
@@ -28,6 +30,7 @@ declare global {
       authService: AuthService;
       categoryService: CategoryService;
       reminderService: ReminderService;
+      userService: UserService;
       middlewareProcessed: boolean;
     }
   }
