+++ server/src/routes/auth.ts
@@ -1,6 +1,10 @@
 import express from 'express';
 import { Middleware } from '../middlewares/middleware';
 import { asyncHandler, errorHandler } from '../middlewares/errorHandler';
+import jwt from 'jsonwebtoken';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 const router = express.Router();
 router.use(Middleware);
@@ -19,6 +23,8 @@ router.post(
   })
 );
 
+const JWT_SECRET = process.env.JWT_SECRET_KEY;
+
 router.post(
   '/login',
   asyncHandler(async (req: express.Request, res: express.Response) => {
@@ -32,65 +38,23 @@ router.post(
 
     const { password, ...userWithoutPassword } = result.user;
 
+    if (!JWT_SECRET) {
+      throw new Error('JWT_SECRET is not defined');
+    }
+
+    const token = jwt.sign(
+      {
+        type: 'JWT',
+        name: userWithoutPassword.name,
+      },
+      JWT_SECRET
+    );
+
     res.json({
       success: true,
       message: '로그인에 성공했습니다.',
+      token: token,
       user: userWithoutPassword,
-    });
-  })
-);
-
-router.get(
-  '/user/:id',
-  asyncHandler(async (req: express.Request, res: express.Response) => {
-    const result = await req.authService.getUser(Number(req.params.id));
-
-    if (!result || result.length === 0) {
-      return res.status(404).json({ success: false, message: '사용자를 찾을 수 없습니다.' });
-    }
-    res.json({
-      success: true,
-      user: {
-        id: result[0].id,
-        name: result[0].name,
-        todos: result[0].completed_todos,
-      },
-    });
-  })
-);
-
-// 로그아웃
-// req.body - id
-router.post(
-  '/logout/',
-  asyncHandler(async (req: express.Request, res: express.Response) => {
-    const { id } = req.body;
-
-    await req.authService.logout(Number(id));
-
-    return res.json({
-      success: true,
-      message: '로그아웃 되었습니다.',
-    });
-  })
-);
-
-// 회원 탈퇴
-router.delete(
-  '/withdrawal/:id',
-  asyncHandler(async (req: express.Request, res: express.Response) => {
-    const result = await req.authService.withdrawal(Number(req.params.id));
-
-    if (!result) {
-      return res.status(404).json({
-        success: false,
-        message: '일치하는 데이터가 없습니다.',
-      });
-    }
-
-    return res.json({
-      success: true,
-      message: '탈퇴 되었습니다.',
     });
   })
 );
